requirements.txt
accelerate==1.0.1
diffusers==0.31.0
facexlib==0.3.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
insightface==0.7.3
numpy
onnxruntime
opencv-python==4.11.0.86
pillow
pillow-avif-plugin==1.5.0
pillow-heif==0.21.0
sentencepiece
torch==2.6.0
torchvision
transformers==4.48.0
peft==0.14.0
pydantic<2